requirements.txt
Flask==3.0.0
paddleocr>=2.7.0
paddlepaddle==2.6.0
Pillow>=10.0.0
numpy>=1.26.0
pypdfium2>=4.0.0
opencv-python-headless>=4.8.0

